Rename wrong acronym to full name
