+++ tests/odwr/test_e2e.py
@@ -16,10 +16,10 @@ import requests
 from userCode import definitions
 from userCode.ontology import ONTOLOGY_MAPPING
 from userCode.env import API_BACKEND_URL
-from userCode.odwr.dag import all_metadata, post_station, sta_station
+from userCode.OregonWaterResourcesDepartment.dag import all_metadata, post_station, sta_station
 from userCode.helper_classes import get_datastream_time_range, MockValues
-from userCode.odwr.lib import assert_no_observations_with_same_iotid_in_first_page
-from userCode.odwr.types import StationData
+from userCode.OregonWaterResourcesDepartment.lib import assert_no_observations_with_same_iotid_in_first_page
+from userCode.OregonWaterResourcesDepartment.types import StationData
 from userCode.util import to_oregon_datetime
